requirements.txt
streamlit>=1.24.0
pdfplumber>=0.9.0
python-docx>=0.8.11
docx2txt>=0.8
spacy>=3.6.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
fpdf2>=2.7.0
plotly>=5.15.0
matplotlib>=3.7.0
wordcloud>=1.9.0
streamlit-authenticator>=0.2.3
psycopg2-binary>=2.9.0
requests>=2.31.0
Pillow>=10.0.0
rapidfuzz>=3.0.0
tenacity>=8.2.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
bcrypt>=4.0.0